requirements.txt
#!/bin/bash

djangorestframework
django-cors-headers
psycopg2-binary